refactor: optimize home screen
